[ALPS05927682] mdla: fix coverity issue
Resolve unused value and negative offset of an array

MTK-Commit-Id: 774ad0a73994bcd1536f822acb59c38dff41f8af

CR-Id: ALPS05927682
Feature: MDLA Platform

+++ drivers/misc/mediatek/apusys/mdla/interface/mdla_cmd_proc_v2_0.c
@@ -47,7 +47,7 @@ static void mdla_cmd_prepare_v2_0(struct mdla_run_cmd *cd,
 
 	mdla_cmd_debug(
 		"%s: kva=0x%llx(0x%llx+0x%x) mva=0x%08x cnt=%u sz=0x%x\n",
-		__func__, (u64)ce->kva, ce->cmdbuf->kva, cd->offset, ce->mva,
+		__func__, (u64)ce->kva, (u64)ce->cmdbuf->kva, cd->offset, ce->mva,
 		ce->count, ce->cmdbuf->size);
 }
